+++ src/DbTests/SQL/TableDtoWithDifferentColumnOrder.sql
@@ -1,0 +1,1 @@
+﻿SELECT TableCol2, TableCol1 FROM dbo.Table1